murfai/requirements.txt
deepgram-sdk>=3.0.0
requests>=2.31.0
python-dotenv>=1.0.0
vaderSentiment>=3.3.2
pyaudio>=0.2.14
websockets>=12.0
aiohttp>=3.9.0
asyncio-mqtt>=0.16.0
python-dateutil>=2.8.2
huggingface-hub>=0.20.0
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

